Removed copyrighted mathematics questions in favor of public and released test questions. Added basic help information for non-admin commands explaining use cases Some code quality improvements - minor and functionless. Separated the right triangle drawing solution into a new file for further expansion Added spares TODO statements to help with organization

+++ start.py
@@ -1,20 +1,17 @@
+import sqlite3
+from discord.ext import commands
+from sympy.parsing.sympy_parser import parse_expr
 from Math_Modules import trigonometric as tr
 from Math_Modules import polynomials as pn
 from Ranking import Ranking as r
 import censor
 import secrets
-from discord.ext import commands
-from sympy.parsing.sympy_parser import parse_expr
-from datetime import datetime
+
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - #
 # Matthew Byrd
 # 2016 - 2018
 # A Discord Bot designed for educational use, to inspire people to learn mathematics more so than now.
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - #
-
-# TO DO LIST #
-# https://cdn.discordapp.com/attachments/272896747385258004/313147123656949760/416b412.gif #
-
 
 description = '''A Magical Bot Used To Do More Than Magical Things'''
 napier = commands.Bot(command_prefix=">", description=description, pm=True)
@@ -27,13 +24,12 @@ async def on_ready():
           "{}\n"
           "{}\n"
           "-----".format(napier.user.name, napier.user.id))
-    start_time = datetime.time
 
-
-########################################################################################################################
+################################################TRIGONOMETRY############################################################
 
 # Trigonometric Functions #
 # Uses Module Trigonometric (tr) #
+
 
 @napier.command(pass_context=True)
 async def sin(ctx):
@@ -57,7 +53,7 @@ async def tan(ctx):
 
 
 @napier.command(pass_context=True)
-async def solve_triangle(ctx):
+async def solve_right_triangle(ctx):
     """Solves a right triangle, given that the right triangle is possible.
     >solve_triangle angle(deg) | angle(deg) | leg | leg | hypo"""
     await napier.say("I am on it!")
@@ -76,13 +72,7 @@ async def solve_triangle(ctx):
     if not infinite: await napier.send_file(ctx.message.channel, 'image.svg')
     if infinite: await napier.say("The triangle has infinitely many solutions")
 
-# End of Trigonometric Functions #
-
-########################################################################################################################
-
-# Polynomial Manipulation #
-# Uses Polynomials module (pn) #
-
+########################################POLYNOMIAL MANIPULATION#########################################################
 
 @napier.command(pass_context=True)
 async def factor(ctx):
@@ -109,7 +99,7 @@ async def simplify(ctx):
 @napier.command(pass_context=True)
 async def roots(ctx):
     """finds all real roots of a given function.
-    >roots x**2 + 3*x^2...etc Multiplication symbol is required (3x != 3*x)"""
+    >roots x**2 + 3*x^2...etc. Multiplication symbol is required (3x != 3*x)"""
     formatted = ctx.message.content.replace(">roots", "")
     try: await napier.say("```{}```".format(pn.get_roots(formatted)))
 
@@ -117,29 +107,36 @@ async def roots(ctx):
         await napier.say("The given function does not make sense to me. 0_0")
         print("wrong {}".format(e))
 
-
-# End of Manipulation #
-
-########################################################################################################################
-
-# Conic Sections #
+#######################################CONIC SECTIONS###################################################################
 
 #@bot.command(pass_context=True)
 #async def conic_section_simplification(ctx):
 #    pass
 
-
-# End Of Conic Sections #
-
-########################################################################################################################
-
-# Ranking System - WIP #
-# Uses ranking module (r) #
+####################################RANKING SYSTEM######################################################################
 
 @napier.command(pass_context=True)
 async def ranks(ctx):
     """Returns a top 10 list of current math point ranks"""
-    rank_list = r.get_ranks()
+    # Returns a discord-formatted string containing the ranks in order
+    db = sqlite3.connect('Ranking/Rankings.db')
+    adder = db.cursor()
+    adder.execute('SELECT * FROM ranks ORDER BY Points DESC')
+    db_result = adder.fetchall()
+
+    rank_list = ""  # Note not actaully a list, but a string.
+    i = 0
+    while i < 10:
+        try:
+            # First item in the array is ID, Second is Points, Third is True/False answer check.
+            final_row = str(db_result[i]).replace("(", "").replace(")", "").replace(" ", "").split(',')
+            user_name = await napier.get_user_info(final_row[0])  # Loose typed language problems
+            rank_list += "{} is rank {} with {} points\n".format(user_name.display_name, i + 1, final_row[1])
+            i += 1
+        except IndexError:  # Breaks in case there isn't 10 users yet.
+            break
+
+    db.close()
     await napier.say("\nTop 10:\n{}\n".format(rank_list))
 
 
@@ -160,6 +157,7 @@ async def points(ctx):
 
 @napier.command(pass_context=True)
 async def add_points(ctx):
+    """Adds points to the passed user"""
     if r.admin_chk(ctx.message.author.id):
         r.check_id(ctx.message.author.id)
         r.add_points(ctx.message.author.id, int(ctx.message.content.replace(">add_points", "").replace(" ", "")))
@@ -182,7 +180,7 @@ async def answer(ctx):
         await next_question(ctx)
 
     elif response not in cur_aw[0] and guess == 0:
-        r.attempt_answer(ctx.message.author.id)
+        r.set_guess(ctx.message.author.id, 1)
         await napier.say("Sorry <@{}> try again next question!".format(ctx.message.author.id))
 
     elif guess == 1:
@@ -191,6 +189,7 @@ async def answer(ctx):
 
 @napier.command(pass_context=True)
 async def set_guess(ctx):
+    """Sets the guess of the called of the command to the specified number"""
     if r.admin_chk(ctx.message.author.id):
         r.check_id(ctx.message.author.id)
         r.set_guess(ctx.message.author.id, int(ctx.message.content.replace(">set_guess", "").replace(" ", "")))
@@ -199,6 +198,7 @@ async def set_guess(ctx):
 
 @napier.command(pass_context=True)                                                                                                                                                                  
 async def start_challenge(ctx):
+    """Starts the question specified question set"""
     formatted = ctx.message.content.replace(">start_challenge", "").split("|")
     if r.admin_chk(ctx.message.author.id):
         question_set = formatted[0]
@@ -209,28 +209,50 @@ async def start_challenge(ctx):
     else: await napier.say("<@{}> Admin commands are for admins only!".format(ctx.message.author.id))
 
 async def next_question(ctx):
+    """For internal bot use only, moves to the next question in the set"""
     try:
         cur_settings = r.get_question(ctx.message.channel.id)
         cur_channel = napier.get_channel(ctx.message.channel.id)
         await napier.send_file(ctx.message.channel,
-                               r"Math Question Repo\{}\{}\{}".format(cur_channel.name,
-                                                                     cur_settings[1], cur_settings[0]))
+                               r"Math Question Repo\{}\{}\{}".format(cur_channel.name, cur_settings[1], cur_settings[0]))
         r.reset_guess()
     except Exception as e:
         print(e)
         await napier.say("It appears that I've run out of questions!")
 
-# End of Ranking System - WIP #
+#####################################GENERAL FUNCTIONS##################################################################
 
-########################################################################################################################
+@napier.command(pass_context=True)
+async def wot(ctx): await wotloc(ctx)
 
-# Primitive Censorship system, currently underpowered. #
+
+async def wotloc(ctx):
+    """Local version of the command so it can be used both by the bot and user, changes This to ThIs"""
+    try: # Loose solution allowing the bot to pass the literal message or for the command to pass the context message.
+        message = ctx.message.content.replace(">wot", "")
+        channel = ctx.message.channel
+    except AttributeError:
+        message = ctx.content.replace(">wot", "")
+        channel = ctx.channel
+    i = 1
+    formatted_message = ""
+    for items in message:
+        if i % 2 != 0: items = items.upper()
+        else: items = items.lower()
+        formatted_message += items
+        if items == " ": continue
+        i += 1
+    await napier.send_message(channel, formatted_message)
+
 
 
 @napier.event
 async def on_message(message):
-    if message.content in censor.arrBad:
-        await napier.delete_message(message)
+    if message.content in censor.arrBad: await napier.delete_message(message)  # Primitive censorship
+    if message.author.id == "138367390185553920" or message.author.id == '138392070724255744': await wotloc(message)
     await napier.process_commands(message)
+
+
+
 
 napier.run(secrets.token)
